Key Listeners
Added Key listeners to the Unlock and Promote inputs so hitting enter
while in the list promote list will promote the person chosen. While in
the user list for unlock it will switch focus to the reason list, and
while in the reason list it will execute the unlock.

Added initial focus when the Unlock and Promote windows are first
displayed to the user list in both cases.